+++ locust/test/test_locust_class.py
@@ -4,6 +4,7 @@ from locust import InterruptTaskSet, ResponseError
 from locust.core import HttpLocust, Locust, TaskSet, events, task
 from locust.exception import (CatchResponseError, LocustError, RescheduleTask,
                               RescheduleTaskImmediately)
+
 from locust.wait_time import between, constant
 from .testcases import LocustTestCase, WebserverTestCase