+++ locust/test/test_runners.py
@@ -502,7 +502,7 @@ class TestStopTimeout(unittest.TestCase):
 
         class MyTestLocust(Locust):
             task_set = MyTaskSet
-            wait_time = constant(1)
+            wait_time = constant(0)
         
         options = mocked_options()
         runner = LocalLocustRunner([MyTestLocust], options)
@@ -518,12 +518,48 @@ class TestStopTimeout(unittest.TestCase):
         runner.quit()
         self.assertEqual("second", MyTaskSet.state)
         
-        options.stop_timeout = short_time * 2 # allow task iteration to complete, with some margin
+        options.stop_timeout = short_time * 3 # allow task iteration to complete, with some margin
         runner = LocalLocustRunner([MyTestLocust], options)
         runner.start_hatching(1, 1)
         gevent.sleep(short_time)
-        runner.quit()
+        timeout = gevent.Timeout(short_time * 2)
+        timeout.start()
+        try:
+            runner.quit()
+            runner.greenlet.join()
+        except gevent.Timeout:
+            self.fail("Got Timeout exception. Some locusts must have kept runnining after iteration finish")
+        finally:
+            timeout.cancel()
         self.assertEqual("third", MyTaskSet.state)
+
+    def test_stop_timeout_during_on_start(self):
+        short_time = 0.05
+        class MyTaskSet(TaskSet):
+            finished_on_start = False
+            my_task_run = False
+            def on_start(self):
+                gevent.sleep(short_time)
+                MyTaskSet.finished_on_start = True
+
+            @task
+            def my_task(self):
+                MyTaskSet.my_task_run = True
+
+        class MyTestLocust(Locust):
+            task_set = MyTaskSet
+            min_wait = 0
+            max_wait = 0
+        
+        options = mocked_options()
+        options.stop_timeout = short_time
+        runner = LocalLocustRunner([MyTestLocust], options)
+        runner.start_hatching(1, 1)
+        gevent.sleep(short_time / 2)
+        runner.quit()
+
+        self.assertTrue(MyTaskSet.finished_on_start)
+        self.assertFalse(MyTaskSet.my_task_run)
 
     def test_stop_timeout_exit_during_wait(self):
         short_time = 0.05